requirements.txt
apache-airflow==1.10.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-bootstrap3==11.0.0
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-datatables==0.4.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.7.7
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1